refactor: remove redundant
